move dilara's drawer code to base activity
